snow_agent/requirements.txt
google-cloud-aiplatform[adk,agent-engines]>=1.114.0
google-adk>=1.14.1
httpx>=0.27.0
pydantic>=2.11.9
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
cloudpickle>=3.1.1
google-cloud-secret-manager>=2.24.0
